requirements.txt
pandas
numpy
pytz
requests
alpaca-trade-api
